backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.7
django-extensions==3.2.3
django-nose==1.4.7
djangorestframework==3.15.2
djongo==1.3.7
dnspython==2.7.0
dotenv==0.9.9
factory_boy==3.3.3
Faker==37.0.2
iniconfig==2.1.0
mongoengine==0.29.1
nose==1.3.7
packaging==24.2
pluggy==1.5.0
pymongo==3.11.4
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.1
